+++ backend/bin/deal-observer-backend.js
@@ -8,7 +8,7 @@ import '../lib/instrument.js'
 import { createInflux } from '../lib/telemetry.js'
 import { getChainHead, rpcRequest } from '../lib/rpc-service/service.js'
 import { fetchDealWithHighestActivatedEpoch, observeBuiltinActorEvents } from '../lib/deal-observer.js'
-import { findAndSubmitUnsubmittedDeals, submitDealsToSparkApi } from '../lib/spark-api-deal-submitter.js'
+import { findAndSubmitUnsubmittedDeals, submitDealsToSparkApi } from '../lib/spark-api-submit-deals.js'
 
 const {
   INFLUXDB_TOKEN,
