+++ backend/lib/spark-api-submit-deals.js
